Update Cucumber to 4.2.6 (#309)
- Cleaned up the bobcat-bdd template
- Added cucumber Gradle task so tests are now run via CLI, not JUnit
