+++ src/main/resources/templates/bobcat-bdd/src/main/java/__packagePath__/CucumberInjectorSource.java
@@ -12,7 +12,7 @@ public class CucumberInjectorSource implements InjectorSource {
 
 	@Override
 	public Injector getInjector() {
-		return Guice.createInjector(CucumberModules.SCENARIO, new BobcatRunModule());
+		return Guice.createInjector(CucumberModules.createScenarioModule(), new BobcatRunModule());
 	}
 
 }
