--- src/main/resources/templates/bobcat-bdd/src/main/java/__packagePath__/hooks/WindowMaximize.java
@@ -1,19 +1,0 @@
-package @packageName@.hooks;
-
-import com.google.inject.Inject;
-import cucumber.api.java.Before;
-import org.openqa.selenium.WebDriver;
-
-/**
- * A helper class that will maximize window
- */
-public class WindowMaximize {
-
-  @Inject
-  private WebDriver webDriver;
-
-  @Before
-  public void maximize() {
-    webDriver.manage().window().maximize();
-  }
-}
