+++ src/main/resources/templates/bobcat-bdd/src/test/resources/features/example.feature
